+++ src/components/index.js
@@ -1,0 +1,2 @@
+export { default as Gravatar } from './Gravatar';
+export { default as SocialMedia } from './SocialMedia';
