+++ src/sections/AboutMe.js
@@ -1,0 +1,10 @@
+import React from 'react';
+import {Container} from 'bloomer';
+
+export default function AboutMe() {
+  return (
+    <Container>
+
+    </Container>
+  );
+}
