+++ src/sections/Projects.js
@@ -1,0 +1,10 @@
+import React from 'react';
+import {Container} from 'bloomer';
+
+export default function Projects() {
+  return (
+    <Container>
+
+    </Container>
+  );
+}
