+++ src/sections/index.js
@@ -1,0 +1,6 @@
+export { default as AboutMe } from './AboutMe';
+export { default as ContactMe } from './ContactMe';
+export { default as Credits } from './Credits';
+export { default as Projects } from './Projects';
+export { default as Welcome } from './Welcome';
+export { default as WorkExperience } from './WorkExperience';
